fix: handle timeout on SFU WS connections
